writing-assistant/requirements.txt
# Writing Assistant Dependencies

# Core web framework
flask>=3.1.1
flask-cors>=6.0.1
python-dotenv>=1.1.1
sqlalchemy>=2.0.41
httpx>=0.28.1

# LLM API Dependencies
langchain-core>=0.1.0
langchain-community>=0.0.10
pydantic>=2.0.0
requests>=2.25.0
colorama>=0.4.4

# LLM provider packages
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-deepseek>=0.1.0
langchain-ollama>=0.1.0

# Optional dependencies
lmstudio>=0.1.0
streamlit>=1.28.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=1.0.0

# Additional dependencies for writing assistant
markupsafe>=2.1.0
jinja2>=3.1.0
werkzeug>=3.0.0
itsdangerous>=2.1.0
click>=8.1.0 